Added files
train.csv and pred.csv are the data for training and prediction of a logistic regression implemented in the titanic.py.

I have used a grid search in order to find the best C parameter for regularisation, and plotted the ROC curves, confusion matrix, and some metrics as precision, recall, and F1-score to fine-tune the model. 

Managed to get to position 975 with 0.80382 categorisation accuracy.